Created Release Build for Android
The app if fully functional and able to be installed on an Android Device. Tested on Tiramisu (Android 13) on AVD/Android Studio without issues. Tested using the Pixel 6a and 3a devices for responsive design.

Also, the application was installed onto my own device (S20 Ultra) and does not have any issues downloading it or running it.

The next step will testing on IOS device, if possible, and cleaning a bit of the code.